+++ src/inputs/mod.rs
@@ -1,3 +1,4 @@
 pub mod chord_pro;
+pub mod propresenter;
 pub mod songbeamer;
 pub mod ultimate_guitar;